requirements.txt
instagrapi
python-dotenv
pyotp
supabase
schedule
requests
